@@ -4,31 +4,38 @@
 	"private": true,
 	"homepage": "https://flashapp-cz.web.app/",
 	"scripts": {
-		"start": "react-scripts start",
-		"build": "react-scripts build",
-		"test": "react-scripts test",
-		"eject": "react-scripts eject",
-		"deploy": "yarn build && firebase deploy",
-		"build:styles": "postcss src/tailwind.css -o src/styles.css",
-		"prebuild": "yarn build:styles",
-		"prestart": "yarn build:styles"
+		"dev": "next dev",
+		"build": "next build",
+		"start": "next start"
 	},
 	"dependencies": {
+		"@chakra-ui/react": "^1.7.3",
+		"@emotion/react": "^11",
+		"@emotion/styled": "^11",
 		"@material-ui/core": "^4.8.3",
 		"@material-ui/icons": "^4.5.1",
 		"@material-ui/lab": "^4.0.0-alpha.39",
-		"faker": "^4.1.0",
-		"firebase": "^7.6.2",
-		"react": "^16.11.0",
-		"react-dom": "^16.11.0",
-		"react-firebaseui": "^4.0.0",
+		"@mui/icons-material": "^5.2.1",
+		"@mui/lab": "^5.0.0-alpha.59",
+		"@mui/material": "^5.2.3",
+		"@mui/styles": "^5.2.3",
+		"autoprefixer": "^10.4.0",
+		"faker": "^5.5.3",
+		"firebase": "^9.6.1",
+		"framer-motion": "^4",
+		"next": "^12.0.7",
+		"postcss": "^8.4.4",
+		"react": "^17.0.2",
+		"react-dom": "^17.0.2",
+		"react-firebase-hooks": "^4.0.1",
+		"react-firebaseui": "^6.0.0",
 		"react-router-dom": "^5.1.2",
-		"react-scripts": "^3.3.0",
-		"react-spring": "9.0.0-beta.3",
+		"react-spring": "9.3.2",
 		"react-textarea-autosize": "^7.1.2",
-		"react-textfit": "^1.1.0",
+		"react-textfit": "^1.1.1",
 		"react-use-gesture": "5.0.0",
-		"typescript": "^3.7.4"
+		"tailwindcss": "^3.0.0",
+		"typescript": "^4.5.3"
 	},
 	"eslintConfig": {
 		"extends": "react-app"
@@ -46,35 +53,18 @@
 		]
 	},
 	"devDependencies": {
-		"@babel/core": "^7.7.2",
-		"@fullhuman/postcss-purgecss": "^2.0.6",
-		"@types/faker": "^4.1.7",
-		"@types/jest": "24.0.21",
-		"@types/node": "12.12.5",
-		"@types/react": "^16.9.11",
-		"@types/react-dom": "16.9.3",
-		"@types/react-router-dom": "^5.1.2",
-		"autoprefixer": "^9.7.4",
-		"babel-plugin-react-css-modules": "^5.2.6",
-		"create-react-component-folder": "^0.1.30",
-		"gh-pages": "^2.1.1",
+		"@types/faker": "^5.5.9",
+		"@types/jest": "27.0.3",
+		"@types/node": "16.11.12",
+		"@types/react": "^17.0.37",
+		"@types/react-dom": "17.0.11",
 		"husky": "^4.2.1",
-		"postcss-cli": "^7.1.0",
-		"prettier": "^1.19.1",
-		"pretty-quick": "^2.0.1",
-		"tailwindcss": "^1.2.0"
+		"prettier": "^2.5.1",
+		"pretty-quick": "^2.0.1"
 	},
 	"husky": {
 		"hooks": {
 			"pre-commit": "pretty-quick --staged"
 		}
-	},
-	"crcf": [
-		"typescript",
-		"cssmodules",
-		"createindex",
-		"functional",
-		"notest",
-		"uppercase"
-	]
+	}
 }
